rename: CorsConfig.class -> SecurityCorsConfig.class
#15
